remove old carousel replace it with new one and implement it in the home page
