+++ client/src/pages/HomePage.tsx
@@ -76,12 +76,18 @@ const HomePage = () => {
     <PageLayout title="Welcome to Makedle" description={description}>
       <div className="flex flex-col gap-4">
         <div className="rounded">
-          <h2 className="text-xl text-gray-800 p-2">Most Popular Games</h2>
-          <Carousel games={popularGames} loading={loadingPopular} />
+          <Carousel
+            items={popularGames}
+            title="Most Popular Games"
+            loading={loadingPopular}
+          />
         </div>
         <div className="rounded">
-          <h2 className="text-xl text-gray-800 p-2">Recently Added Games</h2>
-          <Carousel games={recentGames} loading={loadingRecent} />
+          <Carousel
+            items={recentGames}
+            title="Most Popular Games"
+            loading={loadingRecent}
+          />
         </div>
         <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
           <div className="border-2 shadow rounded h-48">Small Tile 1</div>
